# Improve performance of Transaction parsing