--- a/NBitcoin.Altcoins/Stratis.cs
+++ b/NBitcoin.Altcoins/Stratis.cs
@@ -358,8 +358,8 @@ namespace NBitcoin.Altcoins
 				TxOutList voutTemp = new TxOutList();
 
 				// Try to read the vin. In case the dummy is there, this will be read as an empty vector.
-				stream.ReadWrite<TxInList, TxIn>(ref vinTemp);
-
+				stream.ReadWrite(ref vinTemp);
+				vinTemp.Transaction = this;
 				var hasNoDummy = (nVersionTemp & NoDummyInput) != 0 && vinTemp.Count == 0;
 				if (witSupported && hasNoDummy)
 					nVersionTemp = nVersionTemp & ~NoDummyInput;
@@ -371,9 +371,9 @@ namespace NBitcoin.Altcoins
 					if (flags != 0)
 					{
 						// Assume we read a dummy and a flag.
-						stream.ReadWrite<TxInList, TxIn>(ref vinTemp);
+						stream.ReadWrite(ref vinTemp);
 						vinTemp.Transaction = this;
-						stream.ReadWrite<TxOutList, TxOut>(ref voutTemp);
+						stream.ReadWrite(ref voutTemp);
 						voutTemp.Transaction = this;
 					}
 					else
@@ -386,7 +386,7 @@ namespace NBitcoin.Altcoins
 				else
 				{
 					// We read a non-empty vin. Assume a normal vout follows.
-					stream.ReadWrite<TxOutList, TxOut>(ref voutTemp);
+					stream.ReadWrite(ref voutTemp);
 					voutTemp.Transaction = this;
 				}
 				if (((flags & 1) != 0) && witSupported)
@@ -433,14 +433,14 @@ namespace NBitcoin.Altcoins
 				{
 					// Use extended format in case witnesses are to be serialized.
 					TxInList vinDummy = new TxInList();
-					stream.ReadWrite<TxInList, TxIn>(ref vinDummy);
+					stream.ReadWrite(ref vinDummy);
 					stream.ReadWrite(ref flags);
 				}
 				TxInList vin = this.Inputs;
-				stream.ReadWrite<TxInList, TxIn>(ref vin);
+				stream.ReadWrite(ref vin);
 				vin.Transaction = this;
 				TxOutList vout = this.Outputs;
-				stream.ReadWrite<TxOutList, TxOut>(ref vout);
+				stream.ReadWrite(ref vout);
 				vout.Transaction = this;
 				if ((flags & 1) != 0)
 				{