… NBitcoin.Bench/Serialization.cs